Add a testing utility for generating content
This is a work-in-progress towards making a utility that can generate
content on a per-test bases.

The idea here is to keep all testing data around in parallel unless
explicitly specified that a clean database is necessary (not-yet-implemented).

To ensure there aren't any content collisions between test data,
we randomly create parts of it. It's the test's job to create the content
it needs prior to running the target behavior.