Add a separate testing configuration
This separates the testing configuration into a different docker-compose
file that extends the base file. This specifically mounts the bin directory
and more importantly assigns a 'testing' database separate from the base
database, 'repository'.

This allows one to have a functional database that maintains its data
regardless of test runs. Previously, we would wipe the database before
a test run.