Pin all versions of dependent packages
This will allow pyup.io to track changes to any dependencies.
We track those changes here because the app has the authority over it's
dependencies. It also more importantly has the ability exercise the
fitness and feasiblity of upgrading those dependencies through the test suite.

We'll need a way to keep this up-to-date as dependencies get added,
but that will be work for another changeset.