Fix "fixture db_tables called directly" in tests
In the pytest 4.1.0 changelog:

> Calling fixtures directly is now always an error instead of a warning.

This is the error on travis:

```
______________ ERROR at setup of test_publish_revision_base_case _______________
Fixture "db_tables" called directly. Fixtures are not meant to be called directly,
but are created automatically when test functions request them as parameters.
See https://docs.pytest.org/en/latest/fixture.html for more information about fixtures, and
https://docs.pytest.org/en/latest/deprecations.html#calling-fixtures-directly about how to update your code.
```

One of the solutions suggested by pytest is to separate the function
that has the fixture decorator and the function that returns the
fixture.  This has been changed in cnx-db so the `db_tables` function
can be imported in cnx-press and elsewhere.